# speed up `M^(Int/2)`