This is a bit faster for normal matrices, but way faster for AbstractMatrix that have a fast sqrt.
```
julia> M2 = rand(2,2);
julia> @btime sqrt(M2);
  1.969 μs (21 allocations: 1.53 KiB)
julia> @btime M2^.5;
  2.235 μs (31 allocations: 2.19 KiB)
julia> M = @SMatrix rand(2,2);

julia> @btime $M^.5;
  2.245 μs (37 allocations: 2.52 KiB)

julia> @btime sqrt($M);
  5.561 ns (0 allocations: 0 bytes)
```